Include reviews for deleted cards, addresses #1
Left outer join is used to find reviews with no matching card or model.
Previously, inner join was used, only finding reviews with a card and
model attached.

Anki stores deleted cards Documents/Anki/name/deleted.txt, which appears
to be a TSV file containing the note ID, model ID, and fields of the
deleted items, but I think this is used only for undo? Because with just
these fields, a review of a deleted card can't be tracked down: the review
table only contains a card ID, which the cards table connects to the note
(and thus the model) or the deck ID.

Tested only with deleted cards/notes, not yet with deleted models, but
commit addresses both.