Hardcoded but abstractable C3.js-based viz
- Card index versus date with size of dot indicating pass rate.
- Histogram of pass rates!

TODO: select which decks to consider (also for CSV export) and which field
to treat as the key. Hardcoded currently to my RTK deck.

TODO: make the crosshairs illuminate the datapoint better. I had to
disable automatic resizing of the hovered-over dot because it wouldn't
remember the dot's custom radius :(.